+++ kanso/Kanso.php
@@ -24,7 +24,7 @@ class Kanso
 	 *
 	 * @var string
 	 */
-	const VERSION = '4.3.0';
+	const VERSION = '4.4.0';
 
 	/**
 	 * Singleton instance of self.